Fix float index

+++ NER_IDCNN_CRF/main.py
@@ -107,7 +107,7 @@ def evaluate(sess, model, name, data, id_to_tag, logger):
 
 def split_sentences(total):
     length = len(total)
-    dev_len = length / 10
+    dev_len = int(length / 10)
     test_len = dev_len
     train_len = length - dev_len - test_len
     print("train_len: {} dev_len: {} test_len: {}".format(train_len, dev_len, test_len))
